requirements.txt
coloredlogs==15.0.1
docuwarp==1.0.2
flatbuffers==24.3.25
humanfriendly==10.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.1
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
protobuf==5.28.3
pyreadline3==3.5.4
sympy==1.13.3
